python: test: do not run test_runner.t under tap-driver.sh
Output for multiple tests from a single process is not valid
TAP, so test_runner.t can't be run under automake's tap-driver.sh.
Unfortunately this means python tests only PASS/FAIL as a whole
under parallel-tests, but for now this is preferred over test
failure. Once #278 is resolved, this change can be reverted.