cleanup/attach: thread/fork/exec safe cleanup
The cleanup code now correctly handles forked children that do not exec, and
cleans up *all* left-over files, including sideflux temp directories and
request and event sockets.  Try a make check and see what's left-over.

This also includes two new commands, list-instances and attach, which do what
each name implies.  The attach command runs either a shell or an arbitrary
command in a flux environment connected to the session or URI specified, if
none is specified it connects to the first one found by list-instances.
List-instances lists all instances at the top level of the nearest enclosing
flux, or all top-level local instances if there is none.  Arguments can cause
it to list all local instances, including those outside the nearest enclosing
flux, and to list the entire tree of instances and sub-instances on the local
system.  It should be reasonably possible to add in UDP discovery if we wire
that up on the flux end as well.