config: path environment variables dedup on push
fixes #311

There are a lot more changes here than I expected to need, the main reason is
that lua paths are horrendously difficult to treat like any other kind of
path, because of the way they treat two contiguous separators as a proxy for
the default path.  As it is, this version de-duplicates all path components
from all inputs, but when it finds an empty path component `<sep><sep>` or
`;;`, it inserts the separator twice into the path and then moves on.  This
works for the lua paths, but disallows any truly empty component.  It was the
best solution I could seem to come up with.

libutil/iterators: adding iterator macros

czmq containers work well, but do not provide foreach macros, these greatly
shorten the code to generate a for-loop to iterate over one of them,
especially the zhash.