config/env: properly handle contiguous separators
After the switch in lua path handling, empty path components returned by
split were not being handled correctly, causing an extra path separator
to be added the first time it was encountered.  This commit adds a test
for this case, and a check to ensure that empty components don't add
extra separators.