python: neutralize error check for kvsitr
This commit adds a method to the FunctionWrapper class to allow users to
easily set the error checker for wrapped functions.  The same can be done with
a number of attribute transformations etc. on the wrapper, but this is a great
deal more self documenting. The new method is also used to mark the
kvsitr_next function as a cannot-fail function, since its end condition is
indistinguishable from failure.