broker: add --socket-directory
Create a top level, unique temporary directory to contain broker
sockets.

When an instance is launched directly by flux-start, the instance
shares this directory and puts all of its ranks' sockets inside it.
This means flux-start needs to create it and share it with each
broker.  The new --socket-directory option enables this.

A simplified directory structure and naming results, e.g.

  /tmp/flux.5wv21M/event
  /tmp/flux.5wv21M/0/broker.pid
  /tmp/flux.5wv21M/0/local
  /tmp/flux.5wv21M/0/req
  /tmp/flux.5wv21M/1/broker.pid
  /tmp/flux.5wv21M/1/local
  /tmp/flux.5wv21M/1/req
  ...

When an instance is being launched via slurm or flux, each
rank creates its own unique temporary directory.  In the cases
where ranks need to find each others' ipc sockets (such as
when an event relay is active), these URI's are shared via PMI
so it is not necessary for them to be computed relative to a known
directory.