modules/content-sophia: libsophia content service
This module provides an implementation of a generic content" service
that uses libsophia for persistent storage of objects by hash.

The content protocol represents hashes as 20-byte raw paylaods and data
similarly as raw payloads to avoid repeated JSON-encoding of keys and
data in what could be a performance critical code path.

Flux API functions for working with the content protocol are provided.
They are essentially wrappers around flux_rpc() and flux_rpc_get() so
they may be used both synchronously and asynchronously.

Sophia database files are stored in the broker scratch directory,
and hence must be removed when the broker shuts down.  (At some point
we may want a way to allow the database to persist after session completion,
but for now the problem at hand is unbounded KVS memory usage).

Load the content-sophia module by default on rank 0.