modules/kvs: add cas module for persist obj store
This module will provide an interface to on-disk content
addressable storage using libsophia.  The message interface
will make it a bit easier to integrate with the current kvs code,
which already has code for dealing with async load/store
request/response messages through the cache.  The master will no longer
be special cased; instead it will send a message this module as its
"upstream" object store.

The protocol represents hashes as 20-byte raw paylaods and data
similarly as raw payloads to avoid repeated JSON-encoding of keys and
data in what could be a performance critical code path.

Sophia database files are stored in the broker scratch directory,
and hence must be removed when the broker shuts down.  (At some point
we may want a way to allow the database to persist after session completion,
but for now the problem at hand is unbounded KVS memory usage).

Load the cas module by default on rank 0.