modules/kvs: extend KVS objstore hierarchy to CAS
Treat the rank 0 object cache the same as the hierarchical cache
on other ranks, except instead of loading/storing objects via
upstream ranks, rank 0 uses the CAS module as its "upstream".

dropcache and temporal expiration of object cache entries are
enabled on rank 0.

This means that some operations that could not stall on rank 0
will now stall while messages are sent/received with the CAS.
The main impact should be on the latency of commit and fence operations.
However, the impact should not exceed by very much the RTT of an RPC
between the kvs module on rank 0 and the cas module on rank 0, we hope
sub-millisecond, as requests are for the most part pipelined.