flux-{start,broker}: propagate argument quoting
Both flux-start and flux-broker have classically taken the user command
through an argz_stringify, causing all arguments to be flattened into a
single argument and requiring odd double quoting to make many commands
work.  This makes flux-start propagate all arguments as-is through to
the next stage.  The flux-broker case is a bit more complicated, because
of the shell default, but is currently set up to default to a shell with
no command, use `${shell} -c <arg>` when there is only one argument, and
directly invoke <arg1>...<argn> when there are multiple.  Basic result
is that commands like `flux start -s 2 flux wreckrun -n 4 sh -c
'./something_per_rank | grep stuff | ...'` do something sensible.  Old
behavior caused that to run sh -c 'sh -c ./something_per_rank ...' which
just set the arguments in the inner sh to the rest of the values, losing
them in the command.