fop: partial implementation of embedded func info
This allows for much more go-like interface conversion, but imposes a
duality between selectors that work with it and those that don't.  I'm
not a fan of this, and now understand why java and objective-c do what
they do, going to implement that, explicitly implemented interfaces +
single inheritence, back on the primary branch.