fop: switch dynamic to explicit interface only
Taking comments to heart, the dynamic resolution stuff has been blown
apart and put back together.  Overall, everything is simpler in this
version, but it has the advantage of, combined with the header
touch-ups, allowing fop to be used without ABI dependence.  The
interface setup would allow the definition of interfaces, and acceptance
of any class that implements that interface, without exposing the
structure of any type inside of flux itself.